# openapi/k8s-admin.yaml
openapi: 3.0.3
info:
  title: K8s Admin API
  version: 1.0.0
  description: Core API contract used by UI, Python/Go clients, and Terraform-driven launch workflows.
servers:
  - url: http://localhost:3000
security:
  - sessionCookie: []
...
paths:
  /api/auth/login:
    post:
      tags: [auth]
      summary: Login and establish session cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: Account temporarily locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/platform-panels:
    get:
      tags: [platform-panels]
      summary: Read panel health summary for selected/active cluster
      parameters:
        - name: clusterId
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: Panel summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformPanelsResponse'
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: No accessible cluster
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: k8s_session
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        mustChangePassword:
          type: boolean
      required: [mustChangePassword]
...
    PanelSummary:
      type: object
      properties:
        namespace:
          type: string
        installed:
          type: boolean
        deploymentCount:
          type: integer
        readyDeploymentCount:
          type: integer
        podCount:
          type: integer
        runningPodCount:
          type: integer
        serviceCount:
          type: integer
        keyDeployments:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              ready:
                type: boolean
              readyReplicas:
                type: integer
              replicas:
                type: integer
    PlatformPanelsResponse:
      type: object
      properties:
        cluster:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            status:
              type: string
        panels:
          type: object
          properties:
            argocd:
              $ref: '#/components/schemas/PanelSummary'
            prometheus:
              allOf:
                - $ref: '#/components/schemas/PanelSummary'
                - type: object
                  properties:
                    hasGrafana:
                      type: boolean
            istio:
              allOf:
                - $ref: '#/components/schemas/PanelSummary'
                - type: object
                  properties:
                    ingressGatewayReady:
                      type: boolean
        error:
          type: string